Separating submods of LNP session layer
